+++ src/test/integration/scala/io/evvo/LocalIslandManagerTest.scala
@@ -1,6 +1,6 @@
 package io.evvo
 
-import io.evvo.agent.defaults.DeleteDominated
+import io.evvo.builtin.deletors.DeleteDominated
 import io.evvo.agent.{CreatorFunction, MutatorFunction}
 import io.evvo.island.population.{Maximize, Objective}
 import io.evvo.island.{EvvoIslandBuilder, LocalIslandManager, StopAfter}
